Move the test bits out of SMSHandler
I've just put them into a helper script so that I can still test the bits that are being added to the SMSHandler class
