+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "mood",
   "version": "1.0.0",
-  "description": "Chatbobt conversation analysis to build a picture of users' mood over time. Project repo for the Chainlink '22 Spring Hackathon project",
+  "description": "Conversation analysis to build a picture of users' mood over time. Project repo for the Chainlink '22 Spring Hackathon project",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
